+++ examples/solid/src/App.tsx
@@ -8,6 +8,7 @@ const App: Component = () => {
     <OverlayScrollbarsComponent
       style={{ width: '222px', height: '222px' }}
       options={{ scrollbars: { theme: 'os-theme-light' } }}
+      defer
     >
       <img src={logo} class={styles.logo} alt="logo" width="333" height="333" />
     </OverlayScrollbarsComponent>